requirements.txt
streamlit
tensorflow==2.10.0
pillow
numpy==1.22.4 
keras==2.10.0
protobuf==3.19.6